+++ pkg/client/userd/commands/commands.go
@@ -2,7 +2,6 @@ package commands
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/spf13/cobra"
 
@@ -40,25 +39,6 @@ func GetCommands(ctx context.Context) cliutil.CommandGroups {
 		)
 		cmd.init(ctx)
 		groups[groupName] = append(group, cmd.cobraCommand(ctx))
-	}
-	return groups
-}
-
-// GetCommandsForLocal will return the same commands as GetCommands but in a non-runnable state that reports
-// the error given. Should be used to build help strings even if it's not possible to connect to the connector daemon.
-func GetCommandsForLocal(ctx context.Context, err error) cliutil.CommandGroups {
-	var groups = cliutil.CommandGroups{}
-	for _, cmd := range commands() {
-		var (
-			groupName = cmd.group()
-			group     = groups[groupName]
-			cc        = cmd.cobraCommand(ctx)
-		)
-		cc.RunE = func(_ *cobra.Command, _ []string) error {
-			// err here will be ErrNoUserDaemon "telepresence user daemon is not running"
-			return fmt.Errorf("unable to run command: %w", err)
-		}
-		groups[groupName] = append(group, cc)
 	}
 	return groups
 }